Fix webview save state bug
The "saveState" method on WebView will persistent
the entire contents of the webview and info about
the backstack. This can rack up megabytes worth of data.
The bundle limit is about 500KB.

The fix is to only save the scroll-y position and
rebuild the webview from other state. This prevents
the bundle size from exceeding the max size of an integer.