Use SQL row number to compute unreads on scroll
Uses the computed SQL row_number() to determine which
articles should be marked as read instead of the "published_at"
date. This solves a bug where articles that were published at the same
time were marked as read on scroll.

By using the row number, the index in the database will match up
with what the user sees resulting in the correct items being marked
as read.