Don't send edit-tag if unchanged (#985)
Addresses unexpected behavior with Miniflux's Google Reader API
where clicking through to a read item will mark it as unread.

This is because the edit-tag endpoint will toggle the item's state
instead of persisting it based on the requested state like "mark read."

The fix is to optimistically lock the changest for unread items and
only mark items as read against the API if they're changed in the local
database.