Restore webview scroll position by percent scroll (#1498)
Given a a view height of 500px and current scroll position
of 100px, the scroll state will save a float value of 0.2.

On rotate, if the viewport height decreases slightly given a different
aspect ratio, say 480px, then the restore position will be 96px, or
20% * 480px.

This replaces the absolute scroll position restore which did not
account for the slight change in max height.